+++ src/router/service_account.rs
@@ -1,14 +1,14 @@
 use crate::{
   core::{
     encryption,
-    error::{Errors, InternalError, INTERNAL_ERROR, NOT_FOUND_ERROR},
+    error::{Errors, InternalError, INTERNAL_ERROR, NOT_ALLOWED_ERROR, NOT_FOUND_ERROR},
   },
   middleware::{json::Json, service_account_authorization::ServiceAccountAuthorization},
   model::{
     service_account::{
       CreateServiceAccount, ServiceAccount, ServiceAccountPagination, UpdateServiceAccount,
     },
-    util::{OffsetAndLimit, Pagination},
+    util::{ApplicationId, OffsetAndLimit, Pagination},
   },
   repository,
 };
@@ -30,6 +30,7 @@ pub const SERVICE_ACCOUNT_TAG: &str = "service-account";
   tags = [SERVICE_ACCOUNT_TAG],
   params(
     OffsetAndLimit,
+    ApplicationId,
   ),
   responses(
     (status = 200, content_type = "application/json", body = ServiceAccountPagination),
@@ -46,10 +47,19 @@ pub async fn all_service_accounts(
     service_account, ..
   }: ServiceAccountAuthorization,
   Query(query): Query<OffsetAndLimit>,
+  Query(application_id): Query<ApplicationId>,
 ) -> impl IntoResponse {
+  let application_id = application_id
+    .application_id
+    .unwrap_or(service_account.application_id);
+  if !service_account.is_admin() && service_account.application_id != application_id {
+    return InternalError::unauthorized()
+      .with_error("view-service-accounts", NOT_ALLOWED_ERROR)
+      .into_response();
+  }
   let rows = match repository::service_account::get_service_accounts(
     &state.pool,
-    service_account.application_id,
+    application_id,
     query.limit,
     query.offset,
   )
@@ -84,6 +94,7 @@ pub async fn all_service_accounts(
   tags = [SERVICE_ACCOUNT_TAG],
   params(
     ("service_account_id" = i64, Path, description = "ServiceAccount ID"),
+    ApplicationId,
   ),
   responses(
     (status = 200, content_type = "application/json", body = ServiceAccount),
@@ -101,10 +112,19 @@ pub async fn get_service_account_by_id(
     service_account, ..
   }: ServiceAccountAuthorization,
   Path(service_account_id): Path<i64>,
+  Query(application_id): Query<ApplicationId>,
 ) -> impl IntoResponse {
+  let application_id = application_id
+    .application_id
+    .unwrap_or(service_account.application_id);
+  if !service_account.is_admin() && service_account.application_id != application_id {
+    return InternalError::unauthorized()
+      .with_error("view-service-accounts", NOT_ALLOWED_ERROR)
+      .into_response();
+  }
   let row = match repository::service_account::get_service_account_by_id(
     &state.pool,
-    service_account.application_id,
+    application_id,
     service_account_id,
   )
   .await
@@ -131,6 +151,9 @@ pub async fn get_service_account_by_id(
   path = "/service-accounts",
   tags = [SERVICE_ACCOUNT_TAG],
   request_body = CreateServiceAccount,
+  params(
+    ApplicationId,
+  ),
   responses(
     (status = 201, content_type = "application/json", body = ServiceAccount),
     (status = 400, content_type = "application/json", body = Errors),
@@ -146,8 +169,23 @@ pub async fn create_service_account(
   ServiceAccountAuthorization {
     service_account, ..
   }: ServiceAccountAuthorization,
+  Query(application_id): Query<ApplicationId>,
   Json(payload): Json<CreateServiceAccount>,
 ) -> impl IntoResponse {
+  let application_id = application_id
+    .application_id
+    .unwrap_or(service_account.application_id);
+  if !service_account.is_admin() && service_account.application_id != application_id {
+    return InternalError::unauthorized()
+      .with_error("view-service-accounts", NOT_ALLOWED_ERROR)
+      .into_response();
+  }
+  let is_admin = payload.admin.unwrap_or(false);
+  if is_admin && !service_account.is_admin() {
+    return InternalError::unauthorized()
+      .with_error("create-admin-service-account", NOT_ALLOWED_ERROR)
+      .into_response();
+  }
   let client_id = payload.client_id.unwrap_or_else(uuid::Uuid::new_v4);
   let client_secret = payload.client_secret.unwrap_or_else(uuid::Uuid::new_v4);
   let encrypted_client_secret =
@@ -162,11 +200,12 @@ pub async fn create_service_account(
     };
   let row = match repository::service_account::create_service_account(
     &state.pool,
-    service_account.application_id,
+    application_id,
     repository::service_account::CreateServiceAccount {
       name: payload.name,
       client_id: client_id.to_string(),
       encrypted_client_secret,
+      admin: is_admin,
     },
   )
   .await
@@ -190,7 +229,8 @@ pub async fn create_service_account(
   tags = [SERVICE_ACCOUNT_TAG],
   request_body = UpdateServiceAccount,
   params(
-    ("service_account_id" = i64, Path, description = "ServiceAccount ID")
+    ("service_account_id" = i64, Path, description = "ServiceAccount ID"),
+    ApplicationId
   ),
   responses(
     (status = 201, content_type = "application/json", body = ServiceAccount),
@@ -208,9 +248,24 @@ pub async fn update_service_account(
     service_account, ..
   }: ServiceAccountAuthorization,
   Path(service_account_id): Path<i64>,
+  Query(application_id): Query<ApplicationId>,
   Json(payload): Json<UpdateServiceAccount>,
 ) -> impl IntoResponse {
+  let application_id = application_id
+    .application_id
+    .unwrap_or(service_account.application_id);
+  if !service_account.is_admin() && service_account.application_id != application_id {
+    return InternalError::unauthorized()
+      .with_error("view-service-accounts", NOT_ALLOWED_ERROR)
+      .into_response();
+  }
   let mut params = repository::service_account::UpdateServiceAccount::default();
+  params.admin = payload.admin;
+  if params.admin == Some(true) && !service_account.is_admin() {
+    return InternalError::unauthorized()
+      .with_error("update-admin-service-account", NOT_ALLOWED_ERROR)
+      .into_response();
+  }
   if let Some(client_id) = payload.client_id {
     params.client_id = Some(client_id.to_string().to_owned());
   }
@@ -229,7 +284,7 @@ pub async fn update_service_account(
   }
   let row = match repository::service_account::update_service_account(
     &state.pool,
-    service_account.application_id,
+    application_id,
     service_account_id,
     params,
   )
@@ -260,7 +315,8 @@ pub async fn update_service_account(
   path = "/service-accounts/{service_account_id}",
   tags = [SERVICE_ACCOUNT_TAG],
   params(
-    ("service_account_id" = i64, Path, description = "ServiceAccount ID")
+    ("service_account_id" = i64, Path, description = "ServiceAccount ID"),
+    ApplicationId
   ),
   responses(
     (status = 204),
@@ -278,10 +334,19 @@ pub async fn delete_service_account(
     service_account, ..
   }: ServiceAccountAuthorization,
   Path(service_account_id): Path<i64>,
+  Query(application_id): Query<ApplicationId>,
 ) -> impl IntoResponse {
+  let application_id = application_id
+    .application_id
+    .unwrap_or(service_account.application_id);
+  if !service_account.is_admin() && service_account.application_id != application_id {
+    return InternalError::unauthorized()
+      .with_error("view-service-accounts", NOT_ALLOWED_ERROR)
+      .into_response();
+  }
   match repository::service_account::delete_service_account(
     &state.pool,
-    service_account.application_id,
+    application_id,
     service_account_id,
   )
   .await
